Reorganize library
